+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 * Fixed Platform.* code which couldn't be used in web.
 * Fixed issues where web couldn't use some of report mode or report handler.
 * Added getInstance method to Catcher.
+* Refactored email handlers. 
+* Added html escape in email handlers.
 
 ## 0.6.1
 * Refactored platforms to fix pub dev platforms score.